BUG: Fix incorrect parsing of Label in batch file
When a column "Label" has been specified in the batch input file, the value should be used for the extraction label.
However, this is required to be an integer, whereas it may be parsed as string. Therefore, add a check to convert to integer if it is a string (and therefore not None).

When extracting in parallel, info messages of each thread are filtered out of the console output. However, info messages from the script, e.g. detailing how long the extraction of a single patient took, should not be filtered out. In the batch script, this filter had the wrong logger name passed to the filter (radiomics.batch instead of radiomics.script).