ENH: Change signature of customizable functions to use `kwargs`
All functions that can be customized in the settings, or call functions that can be customized, now use a signature using `kwargs`, which will contain all customization specified in the `setting` section. Necessary elements are obtained using `get()` at the start of the function.

This enables easier addition of new features (only needed to update validation and the function itself, which saves updating the calls to the function throughout pyradiomics). Additionally, this also allows users to replaces functions in PyRadiomics with their own, so long as it follows the same signature (but now does allow for additional/different customization arguments).