BUG: Fix error in generation of provenance info with misaligned mask
When image and mask do not share an identical geometric space, the general info elements for ROI mean, minimum and maximum are incorrect. In case where the mask size does not match the image size, this even produces an error.

Therefore, do not include the image prior to the image and mask geometry are ensured to be the same (this is only the case for the original mask, as the next step is either resampling (forcing similar geometry) or checkMask (comparing/correcting geometry)).

Additionally, add spacing and size elements of the mask for comparison.