STYL: Update naming of provenance info features
Features providing diagnostic/provenance info are now encoded as diagnostics_<Type>_<Feature>.
Here <Type> is either "Versions", "Configuration", "Image-<specifier>" or "Mask-<specifier>". In the latter 2 cases, the specifier indicates where in the preprocessing pipeline the image or mask is: "original" (as from file), "interpolated" (after resampling, only available if enabled), "corrected" (only mask, if mask correction was needed) and "resegmented" (only mask, if resegmentation is enabled).

Finally <Feature> indicates the element in that particular group, e.g. "Size" for the x, y and z dimensions of the image or mask.