ENH: Calculate Volume based on triangle mesh
In concordance with the IBSI definition of volume, calculate it using the triangle mesh that is also generated for the calculation of surface area.
This yields a better correspondence between the calculated volume and surface area, needed for calculation of several subsequent shape features (e.g. sphericity).

Original formula for calculation is retained as 'new' feature "ApproximateVolume".