STYL: Update shape documentation and feature names
To better reflect computation method, explicitly name the two volume features `MeshVolume` and `VoxelVolume`.
Add documentation specifying the unless otherwise mentioned, features are based on the mesh approximation of the shape.

Only exceptions to this are the `VoxelVolume` and PCA-derived features (`MajorAxisLength`, `MinorAxisLength`,`LeastAxisLength`, `Elongation` and `Flatness`. For these features, it is explicitly stated in the documentation that these do not make use of the mesh approximation, but the segmented voxel(center)s themselves.