BUG: fix building pywavelets from source
Since release 1.0.0 of PyWavelets, python 3.4 is no longer officially supported.
Therefore, no pre-built wheels are available on PyPi. To retain support for Python 3.4 on windows,
it has to be build from source (pulled form PyPi). For this, both Cython and Numpy need to be pre-installed.

For Linux and Mac, 3.4 wheels are available for the latest release.