BUG: Fix a bug in MCC function
Improve the fix added in e15f169.

In case of a flat region, GLCM has shape [1, 1, a], thereby resulting only in 1 eigenvalue (1). To prevent an indexing error, return arbitrary 1 in this case.

In certain cases linalg.eigvals returns a complex number. Cast to float to ensure just the real part is returned.