ENH: Implement thread-safe logging
When extracting features in parallel, conflicts may arise when writing to the log-file, corrupting the records.
To prevent this from happening, use a QueueHandler in combination with a QueueListener to process log records for the log file.

Log messages written to the console are not affected.

Additionally, use dictConfig to configure the logging.

Finally, do not catch the Keyboard and SystemExit errors.