ENH: propagate tracking identification from SEG, if available
TrackingIdentifier and TrackingUniqueIdentifier are required. If not specified,
TrackingUniqueIdentifier will be initialized automatically by dcmqi. If either
one is available in SEG, propagate to link with the measurements.

Related PR for dcmqi will make sure TrackingUniqueIdentifier is always created
for SEG segments. It does not hurt, but will become useful if those segments are
later used for measurements. See https://github.com/QIICR/dcmqi/pull/365.