ENH: Add support for overlapping masks / Slicer Segementation objects
In the new version of Slicer, segmentation is done using Segmentation Objects, which allow for overlapping segmentation. When stored, these `.seg.nrrd` files can be read by SimpleITK as vector images. By extracting the nth index from the vector, a labelmap (3D volume) can be obtained representing the nth segement. To ensure correct functionality in the rest of PyRadiomics, all non-zero in the obtained mask are then set to the label value.