BUG: Fix deleting empty gray levels in GLDM and GLRLM
When extracting voxel-wise, there may be discretised gray levels that are not present in a given kernel. In that case, an error occurred in GLRLM and GLDM, where the ivector (defining the gray levels) contained all gray levels, but the matrices only the gray levels present in the mask.

Fix the bug by removing gray levels using the gray levels coefficient.