ENH: Add CLI for voxel-based extraction
Add a new switch in PyRadiomics commandline entrypoint: `--mode`, `-m`, which takes either `segment` (default) or `voxel`.

Add a new switch in PyRadiomics commandline entrypoint: `--out-dir`, `-od`, which specifies a directory (created if not exists) where featuremaps (mode `voxel`) or intermediate files (mode `segment`) are stored. If not specified, featuremaps are stored in the current working directory and intermediate files are not stored.

Refactor the commandline parsing and running of PyRadiomics.
Allow adding diagnositic info to the output when extracting voxel-based features.

After extracting features in a case, results are split into featuremaps, which are stored as SimpleITK images and non-features, which are processed as the output of segment-based extraction.