BUG: Fix bug in firstorder voxel-based calculation
In firstorder, there was an incorrect call in `_initCalculation`, calling the `_initSegmentBasedCalculation` function. When extracting features in segment-based mode, this did not change the output, it only represented a redundant call.
However, when features are extracted in voxel-based mode, the coordinates used to build up the target voxel array was incorrectly reset to all voxels inside the ROI, instead of just the voxels in the current kernel mask. This caused all voxels in the parameter map to have the identical value (and equal to the value from segment-based extraction).