ENH: Add C support for by-voxel calculation
Add an extra dimension as the first dimension of each texture matrix, which will correspond to the matrix for the segment (segment-based, size 1) or the specific kernel identified by the voxel index in the passed array (voxel-based).
This will allow a more vectorized calculation of kernels when processing in by-voxel mode.

Additionally, update the python formulas to allow for this additional dimension by adding axis specifications to all reducing operations and updating defined axes to shift by +1.

Use numpy.squeeze to convert the returned result in segment-based calculation from a 1-D array with size 1 to a scalar float value.