ENH: Implement vectorized voxel-based calculation
Implement vectorized calculation of voxel kernels and remove need of kernel generation for voxel-based calculation of texture matrices.
For the latter, define kernels by updating the bounding box for calculation on the fly in the C extension. The mask array is then only utilized to either mask the kernels (`maskedKernel=True`, maskArray elements true where segmented) or no mask, in which case the maskArray is set to `numpy.ones` of `shape(imageArray)`. This reduces both the number of python-based loops, as the iteration time in the C extension (only iterates over the kernel for each voxel, not the entire image).

Vectorize all feature formulas (exception: shape and shape2D features)

Update firstorder to work with vectorized calculation (add dimension to `targetVoxelArray`, add `discretizedTargetVoxelArray`.

Add voxelSetting parameter `voxelBatch`, which controls the maximum number of voxel calculated in a vectorized way (allows specification of less voxels in 1 batch to conserve memory). Default setting of -1 means all voxels are calculated in one batch

Additionally, remove `self.matrix`, as this is reference equal to `self.imageArray` when applying binning.