ENH: Allow different type for parameter file path specification
A file location may also be passed as e.g. an instance of pathlib WindowsPath, which is not recognized as a string, but does represent a valid file path. Therefore, change the check to see if the string representation of the passed object resolves to an existing file, and if so, use that file as the parameter file.