BUG: Fix resampling error when spacing is equal to resampledPixelSpacing
When spacing is already correct, no resampling is required. However, if the size does not match, or the origin or direction is different, there can still be a geometry mismatch.

Therefore, if the original image spacing is equal to the resampledPixelSpacing, apply resampling to just the mask, and then apply pre-crop.