STYL: Prevent warning C4133 (incompatible types)
In wrapper for calculation of texture matrices, pointers to the voxels array (3 datatypes) need to be initialized to NULL, which was done using a single line. However, this throws an incompatible datatypes warning when compiling, even though the code runs fine. Perform assignment on separate lines to prevent this warning.