BUG: Use local scope copy of settings
When extracting features in parallel, the extractor instance is shared across workers.
To prevent customized updates of `label` and `label_channel` from affecting different cases, use a local copy of `self.settings`.

Then, perform per-case custom updates of settings on this local copy, and pass it to the required functions.