TEST: Fix bug in loading image/mask during testing
In commit 43578f7, use of class instance scope settings was changed to local scope settings allow correct parallel extraction using a shared instance of the extractor.
However, this also meant that the settings had to be passed to the function that loads the image and mask. While this was done so for the call to this function from inside RadiomicsFeatureExtractor, this was not done so for the call to this function in TestUtils, which is the workflow for loading the image/mask during testing. This resulted in ignoring settings during loading, which meant no normalization or resampling was applied. This only affected the testing, and not the regular flow of PyRadiomics.