BUG: Fail gracefully when grayvalue = 0
"0" is an invalid value for gray level and causes an index error in the calculation of texture matrices (e.g. for GLSZM, correct index in matrix is calculated as (i, j) = (grayvalue - 1, region size), which in case of gray value 0 results in index -1, which is invalid).

In some case, this resulted in a graceful error where an IndexError is raised, but this is most likely because the datatype used for the index (`size_t`) does not allow for negative indices on some systems and therefore results in an overflow, which on the subsequent check is then indeed larger than the maximum allowed index (which is checked). In other cases, `size_t` datatype does allow for negative values and therefore passes the maximum index check and subsequently fails with a Process terminated.

Add an additional check to ensure the gray level is > 0, otherwise, fail the C Extension gracefully. Non-gray level properties (i.e. run length, zone size, dependence counts) are calculated in the C extension and ensured to be >= 0.