COMP: Package builds are very sensitive to environment
The pyradiomics package is very sensitive to new envirnments
for building and testing.  If the environment does not closely
match the one used for building, either SimpleITK or the compiled
modules in radiomics/src cause test failures (i.e. segmentation faults).

ubuntu-latest (24.04) and windows-latest (2022) both fail to build
the simpleitk wheel as part of the testing process.